# frontend/trufotbot-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "trufotbot",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/patients": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "List all patients",
        "operationId": "patients_list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Create a patient",
        "operationId": "patients_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Patient created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientCreateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/patients/{id}": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get a patient",
        "operationId": "patients_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Patient ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientGetResponse"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found"
          }
        }
      },
      "put": {
        "tags": [
          "patients"
        ],
        "summary": "Update a patient",
        "operationId": "patients_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Patient ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Patient updated successfully"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      },
      "delete": {
        "tags": [
          "patients"
        ],
        "summary": "Delete a patient",
        "operationId": "patients_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Patient ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient deleted successfully"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      }
    },
    "/api/patients/{id}/test-notification": {
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Send a test notification for this patient",
        "operationId": "patients_test_notification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Patient ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test notification sent successfully"
          },
          "400": {
            "description": "Error from Telegram"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AvailableDose": {
        "type": "object",
        "required": [
          "time"
        ],
        "properties": {
          "quantity": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "None means \"we don't know the amount\""
          },
          "time": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MedicationId": {
        "type": "integer",
        "format": "int32"
      },
      "MedicationSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "next_doses"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/MedicationId"
          },
          "inventory": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "last_taken_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "next_doses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AvailableDose"
            }
          }
        }
      },
      "Patient": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "PatientCreateRequest": {
        "type": "object",
        "description": "Request for POST `/api/patients`, PUT `/api/patients/{patient_id}`",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "telegram_group_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
      "PatientCreateResponse": {
        "type": "object",
        "description": "Response for POST `/api/patients/`.`",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/PatientId"
          }
        }
      },
...
      "PatientGetResponse": {
        "type": "object",
        "description": "Response for GET `/api/patients/{patient_id}`.",
        "required": [
          "name",
          "medications"
        ],
        "properties": {
          "medications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicationSummary"
            }
          },
          "name": {
            "type": "string"
          },
          "telegram_group_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
      "PatientId": {
        "type": "integer",
        "format": "int32"
      },
...
    }
  },
...
}